+++ pages/administratorOverview/administratorOverview.sql
@@ -26,8 +26,8 @@ select_networks AS (
         coalesce(
             jsonb_agg(jsonb_build_object(
                     'network', n.network,
-                    'start_date', coalesce(format_date_full_compact(lower(n.during), 'America/Chicago'), '—'),
-                    'end_date', coalesce(format_date_full_compact(upper(n.during), 'America/Chicago'), '—'),
+                    'start_date', coalesce(format_date_full_compact(lower(n.during), config_select('display_timezone')), '—'),
+                    'end_date', coalesce(format_date_full_compact(upper(n.during), config_select('display_timezone')), '—'),
                     'location', n.location,
                     'purpose', n.purpose
                 ) ORDER BY n.during, n.network),