+++ sprocs/index.js
@@ -71,6 +71,8 @@ module.exports = {
             'submissions_insert.sql',
             'assessment_instances_update.sql',
             'assessment_instances_update_homework.sql',
+            'ip_to_mode.sql',
+            'config_select.sql',
         ], function(filename, callback) {
             logger.verbose('Loading ' + filename);
             fs.readFile(path.join(__dirname, filename), 'utf8', function(err, sql) {