+++ pages/studentInstanceQuestionHomework/studentInstanceQuestionHomework.js
@@ -190,7 +190,8 @@ function processGet(req, res, variant_id, callback) {
                 trueAnswer: res.locals.showTrueAnswer ? res.locals.variant.true_answer : null,
                 submissions : res.locals.showSubmissions ? res.locals.submissions : null,
             });
-            res.locals.questionJsonBase64 = (new Buffer(questionJson)).toString('base64');
+            var encodedJson = encodeURIComponent(questionJson);
+            res.locals.questionJsonBase64 = (new Buffer(encodedJson)).toString('base64');
             res.locals.video = null;
             callback(null);
         },