+++ tests/testHomework.sql
@@ -82,3 +82,14 @@ SET
     max_points = 13
 WHERE
     tid = 'homework1';
+
+-- BLOCK select_last_job_sequence
+SELECT *
+FROM job_sequences
+ORDER BY start_date DESC
+LIMIT 1;
+
+-- BLOCK select_job_sequence
+SELECT *
+FROM job_sequences
+WHERE id = $job_sequence_id;