+++ ChangeLog.md
@@ -1,6 +1,8 @@
 
 # ChangeLog
 
+  * Shift documentation to Read the Docs.
+
 * __2.8.0__ - 2017-05-04
 
   * Add DB storage of exam mode networks.